+++ packages/eslint-config-react/index.js
@@ -26,7 +26,6 @@ module.exports = {
     '@typescript-eslint',
     'react',
     'react-refresh',
-    'unicorn',
     'simple-import-sort',
     'import',
     'unused-imports'