+++ packages/auto-import/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @raipiot-infra/auto-import
 
+## 0.0.15
+
+### Patch Changes
+
+- update
+
 ## 0.0.14
 
 ### Patch Changes