+++ packages/eslint-config-next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @raipiot-infra/eslint-config-next
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @raipiot-infra/eslint-config-base@0.0.6
+
 ## 0.0.7
 
 ### Patch Changes