+++ packages/eslint-config-react-native/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @raipiot-infra/eslint-config-react-native
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @raipiot-infra/eslint-config-base@0.0.6
+
 ## 0.0.6
 
 ### Patch Changes