+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @raipiot-infra/eslint-config-react
 
+## 0.0.9
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @raipiot-infra/eslint-config-base@0.0.6
+
 ## 0.0.8
 
 ### Patch Changes